Better implementation for full encryption/decryption
Use a combination of AES and RSA, generate an AES key for the message. The AES key is used to encrypt the message, and then in turn is encrypted with the public key from the RSA pair and prepended to the resulting data.

AES256 is used, however, the length of the encrypted AES key depends on the RSA block size.