Add an initialiser for modulus/exponent of public key
Also remove the keySize variable, as it seems that there can't be two keys with the same tag, but with different key sizes (i.e using keySize is redundant).

This is also related to exponent/modulus based initialisation, as deriving the keySize would always require splitting the key into its components (key size == modulus in bits).